chore: style functions refactor
